Extend fromDateField, formTable, tableTransferList
[frontend] - Extending the functionality of the formDateField, formTable
tableTransferList components:

formDateField - Added the isRequired property responsible for applying
the appropriate CSS formatting for the required field.

Added mask property that allows to adjust the mask format depending on
the displayed date format, e.g. year ____

Modification of the isValidDate function by adding validations checking
whether the given date is respectively greater or less than the maximum
or minimum date that can be entered. In the case of positive validation
an error message appropriate to the given case is returned.

tableTransferList - Added classes property to allow the default CSS
values of the table wrapper element to be overridden

tableForm - Replaced the numeric and boolean properties by adding a
common Type property (object, text, numeric, boolean, data). This will
enable the correct sort for a given value type.