Create financial plan update in Accountant
[frontend][backend] - Financial update functionality the plans were
created in the Accounting module. The functionality allows you to manage
the update of the financial plan. Correction of the value for items sent
by the coordinator. Approval of the plan and transfer to the
Director module.

Fixed the presentation of assigned values and corrected values in the
Jasper printout when a new item is added when updating the plan.

The functionality includes the GUI on the server side