Fix calculate option value in Assortment Group
[frontend][backend] - The calculation of option values for an assortment
group has been corrected when a part is modified, for which no option
was selected in the previous state. Currently, the option value is
calculated on the basis of all application parts that contain the option
value, except for the currently modified part. The current value of the
modified part is added to the value of the other parts, which adds up to
the total value of the option for the group.