Create a procurement plan update
[frontend] [backend] - The functionality of updating the public
procurement plan has been created in the Coordinator, Public
Procurement, Accountant and Director modules. The functionality allows
you to manage the update of the public procurement plan. Correction of
the item value in the Coordinator module. Approval of the plan and
transfer to subsequent modules in accordance with the approval path.

Approval of the plan update by the Chief Executive will create an update
of the public procurement plan for the institution.

The functionality includes the GUI on the server side