Create plans realization functionality
[frontend][backend][sql] - The functionality of the implementation of
plans allows for recording purchase invoices and contracts with
contractors in the Coordinator's module. Invoice records is responsible
for the implementation of financial and investment plans of
coordinators.

In the case of a financial plan, it also influences the implementation
of the financial plan of the institution.

Registering invoices under the contract allows you to track the value
of the contract.