Create financial and investment plan update
[frontend][backend]- The functionality of updating the financial and
investments plan has been created in the module Coordinator, Accountant
and Director modules. The functionality allows you to manage the update
of the financial or investments plan. Correction of the item value in
the Coordinator module. Approval of the plan and transfer to subsequent
modules in accordance with the approval path.

In the case of a financial plan, sending the plan by the Coordinator
also generates a correction to the financial plan of the institution.
The final approval of the updates of the coordinators' plans takes place
after the institution's financial plan has been approved by Chief.

The functionality includes the GUI on the server side