Add public procurement applications excel export
[frontend][backend] - A functionality has been added that allows the
list of public procurement applications to be exported to Excel.

The orderObject search field has been added to filter the application
list.

Changed the maximum size of characters in the conditions for
conditionsParticipation and orderImportantRecords fields.

Fixed status disabled for fields amountContractAwardedNet, vat  and
button save for application protocol if application estimation type is
different that DO50.

The functionality includes the GUI on the server side.