Modify update financial plan
[frontend][backend][sql] - The functionality of updating the financial
plan has been changed. Currently, individual coordinator plans are
accepted through the full approval path. Then, after approval by the
Director General of the Coordinator's plan, the institution's plan is
created and approved in accordance with the approval path.

The ability to view fulfillment documents for items in the coordinator's
financial and investment plan has been added.

Possibility of exporting fulfillment documents for items in the
coordinator's plan to XLSX has been added.

The functionality of generating cost types has been added

The validation of the entered year was introduced with the creation of
the plan by the coordinator

Validations have been added to check whether the information on items in
the financial or investment plan in applications for public procurement
has been supplemented in the assortment groups