Add Protocol menu in Public Procurement module
[ver 0_7_3_0]
Added the Public Procurement Protocols menu in the Public Procurement
module for managing protocols from the public procurement procedure.

[ver 0_7_2_2]
The calculation of the amount of implementation in the public
procurement plan of the coordinator has been corrected for the DO50 item
when one item from the coordinator's plan is assigned
to several items in the plan of the institution (the situation may occur
at the time of updating the public procurement plan).

[ver 0_7_2_1]
Fixed a filtering error with special characters (unterminated
parenthetical) for the number field in the list of invoices and
contracts.

[ver 0_7_2_0]
Added a print button for plans and updates of coordinators' plans in the
Director module.

Added Economic Director to the institution's financial plan approval
path.

The display of information about the invoice value in the modification
of the invoice header data has been improved.

Added closing the connection to the database when an error occurs when
generating a printout.