Create functionality "Option Values" in module Coordinator - Realization
[ver-0_11_2_2]

Improvement of the automatic change of the VAT value in the protocol for
the application for the award of a public contract based on data from
the assortment group in a situation where there is one assortment group
for two VAT values within one application.

[ver-0_11_3_0]

Adding functionality that enables recording of "option values" for a
contract, invoice and invoice position related to the implementation of
the financial and investment plan item of the Coordinator and
Institution.

Change in the functionality of calculating the value of "realization
from invoices" for the contract object and "value" for the invoice
object. Currently, the value of attributes is stored in the database and
not dynamically calculated when the objects are called.

Setting the default descending sorting by the "sale date" column
for the list of invoices in the Coordinator window - Realization -
Contracts - Contract data

Change in the number of default downloaded contracts to 50 in the
Coordinator - Realization- Contracts window. Improved performance
related to the change in the functionality of calculating the
implementation value for the contract and invoice

[ver-0_11_3_2]

Adding the columns "contract date from" and "contract date to" to the
list of contracts in the Coordinator and Accountant modules.

Extending the number of characters to 1600 for the applicationProtocol
nonCompetitiveOffer field on the public procurement application in the
Coordinator module